requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
